+++ src/components/progress-bar-stack/progress-bar-stack.tsx
@@ -1,0 +1,8 @@
+import { Element } from '@htmlplus/element';
+
+@Element()
+export class ProgressBarStack {
+  render() {
+    return <slot />;
+  }
+}