+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.0] - 0000-00-00
+
+### Added
+
+- The `Progress Bar` component.
+- The `Progress Bar Stack` component.
+
+### Removed
+
+- removed the `Scroll Indicator` component.
+
+### Updated
+
+- Updated the `scroll-detector` reference.
+
 ## [0.9.0] - 2023-05-03
 
 ### Updated